requirements.txt
# ============================================
# HDOWNLOADER - REQUISITOS (ajustado automaticamente)
# ============================================
# Gerado a partir do ambiente `.venv` com `pip freeze`.
# Incluir apenas pacotes instalados para evitar falhas durante
# `pip install -r requirements.txt`.

amqp==5.3.1
asgiref==3.11.1
beautifulsoup4==4.12.3
billiard==4.2.4
celery==5.3.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==42.0.5
daphne==4.1.2
Django==6.0.2
channels==4.1.0
django-crispy-forms==2.5
greenlet==3.3.2
idna==3.11
kombu==5.6.2
packaging==26.0
pdf2image==1.16.3
pdf2docx==0.5.8
pdfminer.six==20231228
pillow==12.1.1
playwright==1.58.0
prompt_toolkit==3.0.52
pycparser==3.0
pyee==13.0.1
psycopg2-binary==2.9.10
pyOpenSSL==24.1.0
pypdf==4.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pywin32==311
redis==5.0.3
requests==2.31.0
setuptools==82.0.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wheel==0.46.3
python-docx==0.8.11
pandas==2.2.0
openpyxl==3.1.2

# ASGI / Channels
channels==4.3.2
channels_redis==4.3.0

# Pacotes opcionais / que exigem compilação nativa no Windows:
# - psycopg2-binary: exige MSVC; instale apenas se tiver Build Tools.
# - lxml: geralmente requer build tools (ou usar rodas pré-compiladas via conda).
# - pymupdf: requer Visual Studio/SDK para compilar MuPDF; prefira conda.
# Se você depender desses pacotes, instale o Microsoft Visual C++ Build Tools
# com o workload "Desktop development with C++" (MSVC v143 + Windows SDK).

# A instalação dos navegadores do Playwright é feita separadamente:
# python -m playwright install chromium
